confinement report
